Use custom ArrayAdapter
Use a custom ArrayAdapter to control which elements of the ListView are
disabled. Still have problems with the display - even though getView is
called for each list item on the screen, in some cases the text color is
set to the color of a disabled item.  It's as if the view were recycled.