Used inflater to in Adapter's getView() method
Instead of calling the parent class' getView the custom adapter now
calls the inflater to create a list item view (if necessary).

The item separator in the list view was also reduced to zero - disabling
the list item seemed to make the separator disappear so might as well
disable separators to start with.